+++ torch_fidelity/fidelity.py
@@ -40,10 +40,10 @@ def main():
                         help='Calculate FID (Frechet Inception Distance)')
     parser.add_argument('-k', '--kid', action='store_true',
                         help='Calculate KID (Kernel Inception Distance)')
+    parser.add_argument('-r', '--prc', action='store_true',
+                        help='Calculate PRC (Precision and Recall)')
     parser.add_argument('-p', '--ppl', action='store_true',
                         help='Calculate PPL (Perceptual Path Length)')
-    parser.add_argument('-r', '--prc', action='store_true',
-                        help='Calculate Precision, Recall')
     parser.add_argument('--feature-extractor', default=DEFAULTS['feature_extractor'], type=str,
                         choices=FEATURE_EXTRACTORS_REGISTRY.keys(),
                         help='Name of the feature extractor')
@@ -54,7 +54,7 @@ def main():
     parser.add_argument('--feature-layer-kid', default=DEFAULTS['feature_layer_kid'], type=str,
                         help='Name of the feature layer to use with KID metric')
     parser.add_argument('--feature-layer-prc', default=DEFAULTS['feature_layer_prc'], type=str,
-                        help='Name of the feature layer to use with Precision, Recall metrics')
+                        help='Name of the feature layer to use with PRC metrics')
     parser.add_argument('--feature-extractor-weights-path',
                         default=DEFAULTS['feature_extractor_weights_path'], type=str,
                         help='Path to feature extractor weights (downloaded if None)')
@@ -63,7 +63,8 @@ def main():
                         choices=FeatureExtractorBase.SUPPORTED_DTYPES,
                         help='dtype to use inside the feature extractor (default if None)')
     parser.add_argument('--feature-extractor-compile', action='store_true',
-                        help='Compile feature extractor (experimental: may have negative effect on metrics precision)')
+                        help='Compile feature extractor (experimental: may have negative effect on metrics numerical '
+                             'precision)')
     parser.add_argument('--isc-splits', default=DEFAULTS['isc_splits'], type=int,
                         help='Number of splits in ISC')
     parser.add_argument('--kid-subsets', default=DEFAULTS['kid_subsets'], type=int,
@@ -96,9 +97,9 @@ def main():
                         choices=list(INTERPOLATION_REGISTRY.keys()),
                         help='Noise interpolation mode in PPL')
     parser.add_argument('--prc-NN-k', default=DEFAULTS['prc_NN_k'], type=int,
-                        help='Number of nearest neighbours for Precision, Recall')
+                        help='Number of nearest neighbours in PRC')
     parser.add_argument('--prc-batch-size', default=DEFAULTS['prc_batch_size'], type=int,
-                        help='Batch size for Precision, Recall')
+                        help='Batch size in PRC')
     parser.add_argument('--no-samples-shuffle', action='store_true',
                         help='Do not perform samples shuffling before computing splits')
     parser.add_argument('--samples-find-deep', action='store_true',